Revert "make boosts work woo (#12)"
This reverts commit 9616f46424363adbc124da52c6728ca047506403.
