+++ internal/apimodule/account/test/accountcreate_test.go
@@ -16,7 +16,7 @@
    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-package account_test
+package account
 
 import (
 	"bytes"
