+++ internal/apimodule/app/test/app_test.go
@@ -16,4 +16,6 @@
    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-package account_test
+package app
+
+// TODO: write tests
