+++ internal/apimodule/auth/auth.go
@@ -32,9 +32,9 @@ import (
 
 const (
 	// AuthSignInPath is the API path for users to sign in through
-	AuthSignInPath = "/auth/sign_in"
+	AuthSignInPath     = "/auth/sign_in"
 	// OauthTokenPath is the API path to use for granting token requests to users with valid credentials
-	OauthTokenPath = "/oauth/token"
+	OauthTokenPath     = "/oauth/token"
 	// OauthAuthorizePath is the API path for authorization requests (eg., authorize this app to act on my behalf as a user)
 	OauthAuthorizePath = "/oauth/authorize"
 )
