+++ internal/apimodule/auth/test/auth_test.go
@@ -16,7 +16,7 @@
    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-package auth_test
+package auth
 
 import (
 	"context"
